+++ pkg/protocols/http/v1/study_service_routes.go
@@ -77,6 +77,7 @@ func (h *HttpEndpoints) AddStudyServiceAdminAPI(rg *gin.RouterGroup) {
 		studyGroup.POST("/:studyKey/remove-member", mw.RequirePayload(), h.studyRemoveMemberHandl)
 		studyGroup.POST("/:studyKey/rules", mw.RequirePayload(), h.saveStudyRulesHandl)
 		studyGroup.POST("/:studyKey/run-rules", mw.RequirePayload(), h.runCustomStudyRulesHandl)
+		studyGroup.POST("/:studyKey/run-rules-for-single-participant", mw.RequirePayload(), h.runCustomStudyRulesForSingleParticipantHandl)
 		studyGroup.POST("/:studyKey/status", mw.RequirePayload(), h.saveStudyStatusHandl)
 		studyGroup.POST("/:studyKey/props", mw.RequirePayload(), h.saveStudyPropsHandl)
 		studyGroup.DELETE("/:studyKey", mw.RequirePayload(), h.deleteStudyHandl)