+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New endpoint to be able to download response export in a JSON format.
 - New endpoints for handling confidential responses (remove for profiles/researcher access).
 - New endpoint for uploading participant file.
+- New endpoint for running custom rules for a single participant.
 
 ### Changed